sum of squares done:
